Reorganized Upgrade Scenarios Tests for 6.3 stubs and more
